FrontDesk/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.22
pydantic==2.9.0
requests==2.32.3
python-dotenv==1.0.1
aiohttp==3.10.5
livekit==0.12.1
livekit-api==0.6.0
pyjwt==2.8.0
